pyproject.toml
[project]
name = "project-scriber"
version = "1.0.1"
authors = [
  { name="SunneV (Wojciech Mariusz Cichoń)", email="wojciech.m.cichon@gmail.com" },
]
description = "An intelligent tool to map, analyze, and compile project source code for LLM context."
readme = "README.md"
requires-python = ">=3.10"
license = { file="LICENSE" }
keywords = ["llm", "code-analysis", "developer-tools", "context-builder", "source-code"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "pathspec",
    "python-dotenv",
    "rich",
    "tiktoken",
    "pyperclip",
    "tomlkit",
    "tomli; python_version < '3.11'",
]

[project.urls]
Homepage = "https://github.com/SunneV/ProjectScriber"
Issues = "https://github.com/SunneV/ProjectScriber/issues"

[project.scripts]
scriber = "scriber.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scriber"]

[tool.pytest.ini_options]
pythonpath = [
    "."
]